+++ src/tree_types.ml
@@ -9,12 +9,11 @@ module TC =
   struct
     (** Atomic pattern constructor. *)
     type t =
-      | Symb of int * string * Files.module_path
-      (** Symbol with its effective arity, name and module path. *)
-      | Abst
-      (** Abstraction (no arguments, patterns are in β-normal form). *)
-      | Vari of string
-      (** A bound variable represented using a name. *)
+      | Symb of Files.module_path * string * int
+      (** Symbol with its module path, name and effective arity. *)
+      | Vari of int
+      (** A bound variable identified by a ({e branch}-wise) unique
+          integer. *)
 
     (** {b NOTE} the effective arity carried by the representation of a symbol
         is specific to a given symbol instance. Indeed, a symbol (in the sense
@@ -25,9 +24,8 @@ module TC =
     (** [pp oc c] prints tree constructor [c] to output channel [o]. *)
     let pp : t pp = fun oc c ->
       match c with
-      | Abst        -> Format.fprintf oc "λ"
-      | Vari(s)     -> Format.pp_print_string oc s
-      | Symb(a,s,_) -> Format.fprintf oc "%s %d-ary" s a
+      | Vari(d)     -> Format.pp_print_int oc d
+      | Symb(_,s,a) -> Format.fprintf oc "%s %d-ary" s a
 
     (** [compare c1 c2] implements a total order on constructors. *)
     let compare : t -> t -> int = Pervasives.compare
@@ -39,14 +37,14 @@ module TCMap = Map.Make(TC)
 (** {3 Decision trees for rewriting} *)
 
 (** Representation of a branching conditions. *)
-type 'term tree_cond =
+type tree_cond =
   | CondNL of int * int
   (** Are the terms at the given indices convertible? We enforce the invariant
       that the first element is a point of reference, which appears in all the
       convertibility conditions involving a given non-linear variable. *)
-  | CondFV of 'term Bindlib.var array * int
-  (** Are the bound variables (which are free at the time of the checking) of
-      the term at the given index in the array? *)
+  | CondFV of int * int array
+  (** Are the (indexed) bound variables (which are free at the time of the
+      checking) of the term at the given index in the array? *)
 
 (** {!b NOTE} that when performing a {!constructor:tree_cond.CondFV} check, we
     are concerned about variables that were bound in the term being reduced
@@ -56,28 +54,28 @@ type 'term tree_cond =
 
 (** Representation of a tree. The definition relies on parameters since module
     {!module:Terms} depends on the current module, and that would thus produce
-    a dependency cycle. However it should be understood that parameter [`term]
-    will only be instantiated with [Terms.term] and parameter [`rhs] will only
-    only be instantiated  with  [(Terms.term_env, Terms.term) Bindlib.mbinder]
-    (i.e., the representation of a RHS). *)
-type ('term, 'rhs) tree =
+    a dependency cycle. However it should be understood that parameter [`rhs]
+    will only be instantiated with
+    [(Terms.term_env, Terms.term) Bindlib.mbinder] (i.e., the representation
+    of a RHS). *)
+type 'rhs tree =
   | Fail
   (** Empty decision tree, used when there are no rewriting rules. *)
-  | Leaf of (int * (int * 'term Bindlib.mvar)) list * 'rhs
+  | Leaf of (int * (int * int array)) list * 'rhs
   (** The value [Leaf(m, rhs)] stores the RHS [rhs] of the rewriting rule that
       can be applied upon reaching the leaf.  The association list [m] is used
       to construct the environment of the RHS. Note that we do not need to use
       a map here since we only need to insert at the head, and iterate over
       the elements of the structure. *)
   | Cond of
-      { ok   : ('term, 'rhs) tree
+      { ok   : 'rhs tree
       (** Branch to follow if the condition is verified. *)
-      ; cond : 'term tree_cond
+      ; cond : tree_cond
       (** The condition to test. *)
-      ; fail : ('term, 'rhs) tree
+      ; fail : 'rhs tree
       (** Branch to follow if the condition is not verified. *) }
   (** Conditional branching according to a condition. *)
-  | Eos of ('term, 'rhs) tree * ('term, 'rhs) tree
+  | Eos of 'rhs tree * 'rhs tree
   (** End of stack node, branches on left tree if the stack is finished, on
       the right if it isn't.  Required when there are rules with a lower arity
       than some other rule above and when {!val:Tree.rule_order} is set. *)
@@ -88,11 +86,11 @@ type ('term, 'rhs) tree =
       ; store : bool
       (** Whether to store the current term. Stored terms might be used in the
           right hand side, are for constraint checks. *)
-      ; children : ('term, 'rhs) tree TCMap.t
+      ; children : 'rhs tree TCMap.t
       (** Subtrees representing the matching of available constructors. *)
-      ; abstraction : ('term Bindlib.var * ('term, 'rhs) tree) option
+      ; abstraction : (int * 'rhs tree) option
       (** Specialisation by an abstraction with the involved free variable. *)
-      ; default : ('term, 'rhs) tree option
+      ; default : 'rhs tree option
       (** When the available patterns contain a wildcard, this subtree is used
           as a last resort (if none of the {!field:children} match). *) }
   (** Arbitrarily branching node allowing to perform switches (a switch is the
@@ -107,7 +105,7 @@ type ('term, 'rhs) tree =
     to leaves in the tree [t], and let [nb_store] be a function mapping a path
     to the number of nodes that have the {!field:store} tag to [true]. We then
     define the capacity [c] of [t] is [c = max{nb_store(p) | p ∈ P}]. *)
-let rec tree_capacity : ('t, 'r) tree -> int = fun tr ->
+let rec tree_capacity : 'r tree -> int = fun tr ->
   match tr with
   | Leaf(_,_)  | Fail   -> 0
   | Eos(l,r)            -> max (tree_capacity l) (tree_capacity r)
@@ -123,7 +121,7 @@ let rec tree_capacity : ('t, 'r) tree -> int = fun tr ->
     the capacity, see {!val:capacity}.  Laziness allows to (sometimes) avoid
     creating several times the same trees when the rules are not given in one
     go. *)
-type ('term, 'rhs) dtree = int Lazy.t * ('term, 'rhs) tree Lazy.t
+type 'rhs dtree = int Lazy.t * 'rhs tree Lazy.t
 
 (** [empty_dtree] is the empty decision tree. *)
-let empty_dtree : ('term, 'rhs) dtree = (lazy 0, lazy Fail)
+let empty_dtree : 'rhs dtree = (lazy 0, lazy Fail)