+++ CHANGES.md
@@ -1,3 +1,12 @@
+#### Trees simplification (2019-12-05)
+
+Simplification of the decision tree structure
+ - trees do not depend on term variables;
+ - tree constructor type depends on generated at compile-time
+   branch-wise unique integral identifiers;
+ - graph output is more consistent: variables are the same in the
+   nodes and the leaves.
+
 ### 1.0 (2018-11-28)
 
 First major release of Lambdapi. It introduces: